ci: automatically update 'latest' tag
This is a temporary fix to update the 'latest' tag while snoots is
still in development. This is needed because until v1.0.0 is released
the latest v1.0.0-dev.x *is* the latest release. However, it seems
that semantic-release does not support pushing to both the 'dev' and
'latest' tags at once, so we have to do it manually.