fix: populate Post.removed when missing (#32)
This patch makes `Post.removed` always been a boolean (never undefined)
by trying to figure out whether or not a post has in fact been removed.
It does this by checking if `removal_reason`, `removed_by`, or
`removed_by_category` are non-empty. If they are, it assumes that the
post has been removed.

This fixes #32.