feat: replace `null` with `undefined`
This is prep for enabling unicorn/no-null, but it's also good practice
and resulted in nicer code.

BREAKING CHANGE: All internal usages of `null` have been replaced with
`undefined`. This includes mapping `null` to `undefined` in Reddit's API
responses. If you were checking any values for existence with `=== null`
you should replace that check with `=== undefined` (or `== null` or
`== undefined`).