feat: add Client.getAuthorizedScopes()
This method allows you to get the currently authorized oAuth scopes.
Note that it will return `undefined` in the following cases:
  - You are using anonymous auth
  - You have not yet called any methods that required oAuth access

The second case is only an issue if you are creating a brand new Client
object from a previously obtained refresh token. Client.fromAuthCode
will set the scopes, but just creating a new Client will not.

Closes #43