feat: make calls to /random go entirely through OAuth
Turns out that calling /random doesn't actually return a random post.
Instead it redirects (302) _to_ a random post. Now this probably would
be fine except that the OAuth redirect goes to www., which has different
rate-limit and authorization rules. This was messing with the rate limit
calculations, and is just not a great implementation.

To work around this we now do the following:

1. Don't automatically follow redirects.
2. When we encounter a redirect, return some data that the caller can
   use to determine what to do.
3. When calling '/random', extract the post ID from the redirect URI,
   and fetch the post directly, using the OAuth endpoint.

This new implementation has the benefit of staying entirely within the
OAuth world while also being more explicit about what it actually does.